Move incoming_message models
