+++ botx/__init__.py
@@ -1,9 +1,9 @@
-from botx.bot.api.commands.accepted_response import build_accepted_response
-from botx.bot.api.commands.bot_disabled_response import (
+from botx.bot.api.exceptions import UnsupportedBotAPIVersionError
+from botx.bot.api.responses.bot_disabled import (
     BotAPIBotDisabledResponse,
     build_bot_disabled_response,
 )
-from botx.bot.api.exceptions import UnsupportedBotAPIVersionError
+from botx.bot.api.responses.command_accepted import build_accepted_response
 from botx.bot.bot import Bot
 from botx.bot.exceptions import (
     AnswerDestinationLookupError,
@@ -15,12 +15,6 @@ from botx.bot.handler_collector import HandlerCollector
 from botx.bot.models.bot_account import BotAccount
 from botx.bot.models.commands.chat import Chat
 from botx.bot.models.commands.enums import ClientPlatforms, MentionTypes, UserKinds
-from botx.bot.models.commands.incoming_message import (
-    ExpressApp,
-    IncomingMessage,
-    UserDevice,
-    UserEventSender,
-)
 from botx.bot.models.method_callbacks import BotAPIMethodFailedCallback
 from botx.bot.models.outgoing_attachment import OutgoingAttachment
 from botx.bot.models.status.bot_menu import BotMenu
@@ -56,12 +50,22 @@ from botx.client.notifications_api.exceptions import (
 from botx.client.notifications_api.markup import BubbleMarkup, Button, KeyboardMarkup
 from botx.client.users_api.exceptions import UserNotFoundError
 from botx.models.message.entities import Mention
+from botx.models.message.incoming_message import (
+    ExpressApp,
+    IncomingMessage,
+    UserDevice,
+    UserEventSender,
+)
 from botx.models.system_events.added_to_chat import AddedToChatEvent
 from botx.models.system_events.chat_created import ChatCreatedEvent, ChatCreatedMember
 from botx.models.system_events.deleted_from_chat import DeletedFromChatEvent
 from botx.shared_models.chat_types import ChatTypes
 
 __all__ = (
+    "IncomingMessage",
+    "ExpressApp",
+    "UserDevice",
+    "UserEventSender",
     "AnswerDestinationLookupError",
     "AddedToChatEvent",
     "DeletedFromChatEvent",
@@ -86,14 +90,12 @@ __all__ = (
     "ChatNotFoundError",
     "ChatTypes",
     "ClientPlatforms",
-    "ExpressApp",
     "FileDeletedError",
     "FileMetadataNotFound",
     "FinalRecipientsListEmptyError",
     "FinalRecipientsListEmptyError",
     "HandlerCollector",
     "HandlerNotFoundError",
-    "IncomingMessage",
     "InvalidBotAccountError",
     "InvalidBotXResponsePayloadError",
     "InvalidBotXStatusCodeError",
@@ -108,8 +110,6 @@ __all__ = (
     "StealthModeDisabledError",
     "UnknownBotAccountError",
     "UnsupportedBotAPIVersionError",
-    "UserDevice",
-    "UserEventSender",
     "UserKinds",
     "UserNotFoundError",
     "build_accepted_response",
