+++ botx/bot/api/responses/bot_disabled.py
