+++ botx/bot/handler_collector.py
@@ -14,12 +14,12 @@ from botx.bot.handler import (
     VisibleCommandHandler,
     VisibleFunc,
 )
-from botx.bot.models.commands.commands import BotCommand, SystemEvent
-from botx.bot.models.commands.incoming_message import IncomingMessage
 from botx.bot.models.status.bot_menu import BotMenu
 from botx.bot.models.status.recipient import StatusRecipient
 from botx.converters import optional_sequence_to_list
 from botx.logger import logger
+from botx.models.commands import BotCommand, SystemEvent
+from botx.models.message.incoming_message import IncomingMessage
 from botx.models.system_events.added_to_chat import AddedToChatEvent
 from botx.models.system_events.chat_created import ChatCreatedEvent
 from botx.models.system_events.deleted_from_chat import DeletedFromChatEvent
