+++ botx/models/base_command.py
@@ -1,3 +1,4 @@
+from dataclasses import dataclass
 from typing import Any, Dict, Optional
 from uuid import UUID
 
@@ -69,3 +70,9 @@ class BotAPIBaseCommand(VerifiedPayloadBaseModel):
             return version
 
         raise UnsupportedBotAPIVersionError(version)
+
+
+@dataclass
+class BotCommandBase:
+    bot_id: UUID
+    raw_command: Optional[Dict[str, Any]]
